# cmd/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A GEWIS Rooster maker",
        "title": "GRooster",
        "contact": {},
        "version": "0.1"
    },
    "paths": {
...
        "/roster/": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roster"
                ],
                "summary": "Get all rosters",
                "operationId": "getRosters",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Roster after this date",
                        "name": "afterDate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/RosterResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roster"
                ],
                "summary": "CreateRoster a new roster",
                "operationId": "createRoster",
                "parameters": [
                    {
                        "description": "Roster input",
                        "name": "createParams",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RosterCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Roster"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/roster/answer": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roster"
                ],
                "summary": "Create a new roster shift answer",
                "operationId": "createRosterAnswer",
                "parameters": [
                    {
                        "description": "Roster answer input",
                        "name": "createParams",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RosterAnswerCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/RosterAnswer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/roster/answer/{id}": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roster"
                ],
                "summary": "Updates a roster answer with the new value",
                "operationId": "updateRosterAnswer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Roster Answer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New answer value",
                        "name": "updateParams",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RosterAnswerUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/RosterAnswer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/roster/saved-shift/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roster"
                ],
                "summary": "Get all saved shifts for a specific roster",
                "operationId": "getSavedRoster",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Roster ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Saved Shifts",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/SavedShift"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "SavedShift not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roster"
                ],
                "summary": "Update a specific saved shift",
                "operationId": "updateSavedShift",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "SavedShift ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update data",
                        "name": "updateParams",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SavedShiftUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SavedShift"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "SavedShift not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/roster/shift": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roster"
                ],
                "summary": "Create a new roster shift",
                "operationId": "createRosterShift",
                "parameters": [
                    {
                        "description": "Roster shift input",
                        "name": "createParams",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RosterShiftCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/RosterShift"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/roster/shift/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roster"
                ],
                "summary": "Deletes a roster shift",
                "operationId": "deleteRosterShift",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Roster Answer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/roster/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roster"
                ],
                "summary": "Get the organs rosters",
                "operationId": "getOrganRosters",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Organ ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/RosterResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roster"
                ],
                "summary": "DeleteRoster a roster",
                "operationId": "deleteRoster",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Roster ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roster"
                ],
                "summary": "Update a roster",
                "operationId": "updateRoster",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Roster ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Roster input",
                        "name": "updateParams",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RosterUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Roster"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/roster/{id}/save": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roster"
                ],
                "summary": "Save a specific roster",
                "operationId": "rosterSave",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Roster ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Roster": {
            "description": "Roster info with a gorm model, name, users, shifts and values",
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "organ": {
                    "type": "integer"
                },
                "rosterAnswer": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/RosterAnswer"
                    }
                },
                "rosterShift": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/RosterShift"
                    }
                },
                "saved": {
                    "type": "boolean"
                },
                "updatedAt": {
                    "type": "string"
                },
                "values": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "RosterAnswer": {
            "description": "An answer from a user to a shift.",
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "rosterId": {
                    "type": "integer"
                },
                "rosterShiftId": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "RosterAnswerCreateRequest": {
            "type": "object",
            "properties": {
                "rosterId": {
                    "type": "integer"
                },
                "rosterShiftId": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "RosterAnswerUpdateRequest": {
            "description": "Request to update a certain roster answer",
            "type": "object",
            "properties": {
                "value": {
                    "type": "string"
                }
            }
        },
        "RosterCreateRequest": {
            "description": "Roster create request",
            "type": "object",
            "properties": {
                "date": {
                    "description": "Date the date that this roster will take place",
                    "type": "string"
                },
                "name": {
                    "description": "Name the name of the new roster",
                    "type": "string"
                },
                "organId": {
                    "description": "Organ that the roster belongs to",
                    "type": "integer"
                }
            }
        },
        "RosterResponse": {
            "description": "RosterResponse is the response  from a get roster request",
            "type": "object",
            "properties": {
                "roster": {
                    "$ref": "#/definitions/Roster"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/User"
                    }
                }
            }
        },
...
        "RosterUpdateRequest": {
            "type": "object"
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}